awesomefunctions

+++ src/RGBtorgb.m
@@ -1,7 +1,7 @@
 function output = RGBtorgb( RGB_image )
 % RGBtorgb normalizes a RGB image
     % Perform the normalization R / (R+G+B), G / (R+G+B) and B / (R+G+B)
-    output = RGB_image ./ repmat( sum( RGB_image, 3 ), [1,1,3] );
+    output = double( RGB_image ) ./ repmat( sum( double( RGB_image ), 3 ), [1,1,3] );
     % Map all NaN's to 0
     output( isnan( output ) ) = 0;
